refactor: update config files
