+++ server/index.js
@@ -1,0 +1,6 @@
+const app = require('./app');
+const PORT = 3000;
+
+app.listen(PORT, () => {
+    console.log(`Server listening on PORT ${PORT}`)
+});
