+++ src/Root.js
@@ -1,0 +1,12 @@
+import 'babel-polyfill';
+import React from 'react';
+import { hot } from 'react-hot-loader';
+
+
+const Root = () => (
+    <div>
+        Server side rendering
+    </div>
+)
+
+export default hot(module)(Root);
